APIs/fungenerators.com/riddle/1.5/python-aiohttp-client/openapi_server/controllers/private_riddles_controller.py
from typing import List, Dict
from aiohttp import web

from openapi_server import util


async def riddle_delete(request: web.Request, id) -> web.Response:
    """riddle_delete

    Create a random Riddle entry.

    :param id: Riddle ID
    :type id: str

    """
    return web.Response(status=200)


async def riddle_get(request: web.Request, id=None) -> web.Response:
    """riddle_get

    Get a Riddle entry for a given id. Retrieves a riddle question and answer based on the id.

    :param id: ID of the riddle to fetch
    :type id: str

    """
    return web.Response(status=200)


async def riddle_post(request: web.Request, question, category, answer) -> web.Response:
    """riddle_post

    Create a random Riddle entry. Same as &#39;PUT&#39; but can be used when some of the client libraries don&#39;t support &#39;PUT&#39;.

    :param question: Riddle Question
    :type question: str
    :param category: Category of the riddle
    :type category: str
    :param answer: Answer(s) to the riddle question
    :type answer: str

    """
    return web.Response(status=200)


async def riddle_put(request: web.Request, question, category, answer) -> web.Response:
    """riddle_put

    Create a random Riddle entry.

    :param question: Riddle Question
    :type question: str
    :param category: Category of the riddle
    :type category: str
    :param answer: Answer(s) to the riddle question
    :type answer: str

    """
    return web.Response(status=200)
